+++ crates/platform/src/config.rs
@@ -56,6 +56,7 @@ pub struct MoyuConfig {
     pub show_fps: bool,
     #[serde(rename = "enableMSAA")]
     pub enable_msaa: bool,
+    pub enable_mipmaps: bool,
     pub enable_gamepads: bool,
     pub skip_splash: bool,
     /// Custom parameters that can be accessed in the engine.
@@ -82,6 +83,7 @@ impl Default for MoyuConfig {
             background_color: Color::from_html("transparent").unwrap(),
             show_fps: false,
             enable_msaa: false,
+            enable_mipmaps: false,
             enable_gamepads: false,
             skip_splash: false,
             params: String::new(),