+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 1. Implement MSAA support and update rendering pipeline
+2. Implement mipmap support and enable mipmap generation in texture loading
 
 ### Fixed